fix net

+++ app/src/main/java/com/pisces/android/wuha/entity/bean/ServiceProvider.java
@@ -1,6 +1,5 @@
 package com.pisces.android.wuha.entity.bean;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -10,18 +9,18 @@ public class ServiceProvider {
 
 
     /**
-     * Id : 5732485075008993430
-     * Name : 火车北站
-     * ServiceProviderAddress : {"Id":"4947533809332902844","Latitude":104.079853,"Longitude":30.703259,"MainAddressLine":"四川省成都市二环路北三段","Name":"主要地址"}
-     * ServiceProviderContact : {"Id":"5707580329792543729","Name":"主要联系方式","Phone":"87654321","MobilePhone":"10987654321","OrtherContact":"OtherContact"}
-     * ServiceProviderCertificationInfo : {"Id":"4711101685671623188","BusinessLicenseImagePath":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1504340601&di=0c4be8cf6a460c51b9ab0d222eb48dbe&imgtype=jpg&er=1&src=http%3A%2F%2Fcdn.duitang.com%2Fuploads%2Fitem%2F201406%2F27%2F20140627085544_cHFnx.jpeg","LegalPersonInfoImagePath":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503745927460&di=737d4db400e280d58eb23ac26c9eebea&imgtype=0&src=http%3A%2F%2Fpic20.nipic.com%2F20120403%2F7885626_091956683185_2.jpg","IsCertificationInfo":0}
-     * ServiceProviderIntroduction : {"Id":"5498726687780377447","Introduction":"这里是火车北站的简单介绍","Description":"这里是火车北站的详细描述","ImagePath":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1504340745&di=94e8cd756ad398a6f6c4ce132f54253b&imgtype=jpg&er=1&src=http%3A%2F%2Fpic8.nipic.com%2F20100718%2F5401072_123826008131_2.jpg","ServiceProviderBannerPaths":[{"Id":31,"Path":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503746006389&di=0db690e39e4e0570d259395a0d3ed52e&imgtype=0&src=http%3A%2F%2Fpic.qiantucdn.com%2F58pic%2F18%2F09%2F79%2F55d1ed2c1098a_1024.jpg"},{"Id":32,"Path":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503746006378&di=9b3160bc22fb15d3939c593e198079cf&imgtype=0&src=http%3A%2F%2Fimg155.poco.cn%2Fmypoco%2Fmyphoto%2F20090611%2F14%2F12279062200906111412311756765842618_009.jpg"}]}
-     * ServiceProviderServiceCategories : [{"Id":"4757298004609728188","Name":"宠物手术","SubCategories":[{"Id":"4900610231028091476","Name":"小型手术","Price":3800},{"Id":"5037523185200550771","Name":"中型手术","Price":4800},{"Id":"5164803786160131319","Name":"大型手术","Price":5800}]},{"Id":"5226555842452527060","Name":"预诊","SubCategories":[{"Id":"4743673771592853677","Name":"初级诊断","Price":80},{"Id":"5016144334641149622","Name":"VIP服务","Price":120}]},{"Id":"5476616203023100390","Name":"清创","SubCategories":[{"Id":"4804827369523182134","Name":"小面积","Price":1200},{"Id":"4807809450169595507","Name":"大面积","Price":4800}]}]
+     * Id : 5361304320953356438
+     * Name : 成都杜甫草堂博物馆
+     * ServiceProviderAddress : {"Id":"4941700888876585476","Latitude":104.034848,"Longitude":30.666397,"MainAddressLine":"青华路37号","Name":"主要地址"}
+     * ServiceProviderContact : {"Id":"5743370408263164600","Name":"主要联系方式","Phone":"083387654321","MobilePhone":"98764567102","OrtherContact":"OtherContact"}
+     * ServiceProviderCertificationInfo : {"Id":"4981901421985701438","BusinessLicenseImagePath":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1504340601&di=0c4be8cf6a460c51b9ab0d222eb48dbe&imgtype=jpg&er=1&src=http%3A%2F%2Fcdn.duitang.com%2Fuploads%2Fitem%2F201406%2F27%2F20140627085544_cHFnx.jpeg","LegalPersonInfoImagePath":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503745927460&di=737d4db400e280d58eb23ac26c9eebea&imgtype=0&src=http%3A%2F%2Fpic20.nipic.com%2F20120403%2F7885626_091956683185_2.jpg","IsCertificationInfo":0}
+     * ServiceProviderIntroduction : {"Id":"5605410605027501871","Introduction":"这里是成都杜甫草堂博物馆的简单介绍","Description":"这里是成都杜甫草堂博物馆的详细描述","ImagePath":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1504340745&di=94e8cd756ad398a6f6c4ce132f54253b&imgtype=jpg&er=1&src=http%3A%2F%2Fpic8.nipic.com%2F20100718%2F5401072_123826008131_2.jpg","ServiceProviderBannerPaths":[{"Id":33,"Path":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503746006389&di=0db690e39e4e0570d259395a0d3ed52e&imgtype=0&src=http%3A%2F%2Fpic.qiantucdn.com%2F58pic%2F18%2F09%2F79%2F55d1ed2c1098a_1024.jpg"},{"Id":34,"Path":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503746006378&di=9b3160bc22fb15d3939c593e198079cf&imgtype=0&src=http%3A%2F%2Fimg155.poco.cn%2Fmypoco%2Fmyphoto%2F20090611%2F14%2F12279062200906111412311756765842618_009.jpg"}]}
+     * ServiceProviderServiceCategories : [{"Id":"4612561924395694951","Name":"宠物手术","SubCategories":[{"Id":"5156293244959719619","Name":"中型手术","Price":9800},{"Id":"5182525672232723732","Name":"大型手术","Price":15800},{"Id":"5544385201366288029","Name":"小型手术","Price":6800}]},{"Id":"4794391862103782756","Name":"清创","SubCategories":[{"Id":"4837539746106540850","Name":"小面积","Price":3200},{"Id":"4997100854385977888","Name":"大面积","Price":8800}]},{"Id":"4946134811791848949","Name":"诊断服务","SubCategories":[{"Id":"4750213658882502981","Name":"VIP服务","Price":620},{"Id":"5233071810722233772","Name":"初级诊断","Price":80},{"Id":"5638434957439061686","Name":"急诊","Price":120}]}]
      * StartingPrice : 80
-     * ViewingCount : 0
+     * ViewingCount : 3
      * CreateTime : 2017-08-26T00:00:00
      * ServiceProviderType : 1
-     * Distance : 1224.466546031413
+     * Distance : 0
      */
 
     private String Id;
@@ -30,11 +29,11 @@ public class ServiceProvider {
     private ServiceProviderContactBean ServiceProviderContact;
     private ServiceProviderCertificationInfoBean ServiceProviderCertificationInfo;
     private ServiceProviderIntroductionBean ServiceProviderIntroduction;
-    private BigDecimal StartingPrice;
+    private int StartingPrice;
     private int ViewingCount;
     private String CreateTime;
     private int ServiceProviderType;
-    private double Distance;
+    private int Distance;
     private List<ServiceProviderServiceCategoriesBean> ServiceProviderServiceCategories;
 
     public String getId() {
@@ -85,11 +84,11 @@ public class ServiceProvider {
         this.ServiceProviderIntroduction = ServiceProviderIntroduction;
     }
 
-    public BigDecimal getStartingPrice() {
+    public int getStartingPrice() {
         return StartingPrice;
     }
 
-    public void setStartingPrice(BigDecimal StartingPrice) {
+    public void setStartingPrice(int StartingPrice) {
         this.StartingPrice = StartingPrice;
     }
 
@@ -117,11 +116,11 @@ public class ServiceProvider {
         this.ServiceProviderType = ServiceProviderType;
     }
 
-    public double getDistance() {
+    public int getDistance() {
         return Distance;
     }
 
-    public void setDistance(double Distance) {
+    public void setDistance(int Distance) {
         this.Distance = Distance;
     }
 
@@ -135,10 +134,10 @@ public class ServiceProvider {
 
     public static class ServiceProviderAddressBean {
         /**
-         * Id : 4947533809332902844
-         * Latitude : 104.079853
-         * Longitude : 30.703259
-         * MainAddressLine : 四川省成都市二环路北三段
+         * Id : 4941700888876585476
+         * Latitude : 104.034848
+         * Longitude : 30.666397
+         * MainAddressLine : 青华路37号
          * Name : 主要地址
          */
 
@@ -191,10 +190,10 @@ public class ServiceProvider {
 
     public static class ServiceProviderContactBean {
         /**
-         * Id : 5707580329792543729
+         * Id : 5743370408263164600
          * Name : 主要联系方式
-         * Phone : 87654321
-         * MobilePhone : 10987654321
+         * Phone : 083387654321
+         * MobilePhone : 98764567102
          * OrtherContact : OtherContact
          */
 
@@ -247,7 +246,7 @@ public class ServiceProvider {
 
     public static class ServiceProviderCertificationInfoBean {
         /**
-         * Id : 4711101685671623188
+         * Id : 4981901421985701438
          * BusinessLicenseImagePath : https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1504340601&di=0c4be8cf6a460c51b9ab0d222eb48dbe&imgtype=jpg&er=1&src=http%3A%2F%2Fcdn.duitang.com%2Fuploads%2Fitem%2F201406%2F27%2F20140627085544_cHFnx.jpeg
          * LegalPersonInfoImagePath : https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503745927460&di=737d4db400e280d58eb23ac26c9eebea&imgtype=0&src=http%3A%2F%2Fpic20.nipic.com%2F20120403%2F7885626_091956683185_2.jpg
          * IsCertificationInfo : 0
@@ -293,11 +292,11 @@ public class ServiceProvider {
 
     public static class ServiceProviderIntroductionBean {
         /**
-         * Id : 5498726687780377447
-         * Introduction : 这里是火车北站的简单介绍
-         * Description : 这里是火车北站的详细描述
+         * Id : 5605410605027501871
+         * Introduction : 这里是成都杜甫草堂博物馆的简单介绍
+         * Description : 这里是成都杜甫草堂博物馆的详细描述
          * ImagePath : https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1504340745&di=94e8cd756ad398a6f6c4ce132f54253b&imgtype=jpg&er=1&src=http%3A%2F%2Fpic8.nipic.com%2F20100718%2F5401072_123826008131_2.jpg
-         * ServiceProviderBannerPaths : [{"Id":31,"Path":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503746006389&di=0db690e39e4e0570d259395a0d3ed52e&imgtype=0&src=http%3A%2F%2Fpic.qiantucdn.com%2F58pic%2F18%2F09%2F79%2F55d1ed2c1098a_1024.jpg"},{"Id":32,"Path":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503746006378&di=9b3160bc22fb15d3939c593e198079cf&imgtype=0&src=http%3A%2F%2Fimg155.poco.cn%2Fmypoco%2Fmyphoto%2F20090611%2F14%2F12279062200906111412311756765842618_009.jpg"}]
+         * ServiceProviderBannerPaths : [{"Id":33,"Path":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503746006389&di=0db690e39e4e0570d259395a0d3ed52e&imgtype=0&src=http%3A%2F%2Fpic.qiantucdn.com%2F58pic%2F18%2F09%2F79%2F55d1ed2c1098a_1024.jpg"},{"Id":34,"Path":"https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503746006378&di=9b3160bc22fb15d3939c593e198079cf&imgtype=0&src=http%3A%2F%2Fimg155.poco.cn%2Fmypoco%2Fmyphoto%2F20090611%2F14%2F12279062200906111412311756765842618_009.jpg"}]
          */
 
         private String Id;
@@ -348,7 +347,7 @@ public class ServiceProvider {
 
         public static class ServiceProviderBannerPathsBean {
             /**
-             * Id : 31
+             * Id : 33
              * Path : https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1503746006389&di=0db690e39e4e0570d259395a0d3ed52e&imgtype=0&src=http%3A%2F%2Fpic.qiantucdn.com%2F58pic%2F18%2F09%2F79%2F55d1ed2c1098a_1024.jpg
              */
 
@@ -375,9 +374,9 @@ public class ServiceProvider {
 
     public static class ServiceProviderServiceCategoriesBean {
         /**
-         * Id : 4757298004609728188
+         * Id : 4612561924395694951
          * Name : 宠物手术
-         * SubCategories : [{"Id":"4900610231028091476","Name":"小型手术","Price":3800},{"Id":"5037523185200550771","Name":"中型手术","Price":4800},{"Id":"5164803786160131319","Name":"大型手术","Price":5800}]
+         * SubCategories : [{"Id":"5156293244959719619","Name":"中型手术","Price":9800},{"Id":"5182525672232723732","Name":"大型手术","Price":15800},{"Id":"5544385201366288029","Name":"小型手术","Price":6800}]
          */
 
         private String Id;
@@ -410,14 +409,14 @@ public class ServiceProvider {
 
         public static class SubCategoriesBean {
             /**
-             * Id : 4900610231028091476
-             * Name : 小型手术
-             * Price : 3800
+             * Id : 5156293244959719619
+             * Name : 中型手术
+             * Price : 9800
              */
 
             private String Id;
             private String Name;
-            private BigDecimal Price;
+            private int Price;
 
             public String getId() {
                 return Id;
@@ -435,11 +434,11 @@ public class ServiceProvider {
                 this.Name = Name;
             }
 
-            public BigDecimal getPrice() {
+            public int getPrice() {
                 return Price;
             }
 
-            public void setPrice(BigDecimal Price) {
+            public void setPrice(int Price) {
                 this.Price = Price;
             }
         }
